+++ src/strategy/HourlyPricing.java
@@ -1,5 +1,5 @@
 package strategy;
 
 public class HourlyPricing implements PricingStrategy {
-    public double calculatePrice(int duration) { return duration * 10; }
+    public float calculatePrice(int duration) { return duration * 10; }
 }
